updated unaffiliated participants controller
